Fix logic failure related to SPIRV-tools-opt
If the user has specified ENABLE_OPT, he wants to link with the SPIRV-Tools-opt
library. The corresponding cmake target can be provided either by an
external development package for SPIRV-tools or by embedded spirv-tools
sources. In the first case, the corresponding `find_package` call was missing.
In the second case, the detection of this case was insufficient.

The distinction of the states is now done with `ENABLE_OPT AND SPIRV-Tools-opt_FOUND`
for the use of the external package and with `ENABLE_OPT AND spirv-tools_SOURCE_DIR`
in the case of the embedded sources.

The cmake variable `spirv-tools_SOURCE_DIR` should be defined by calling
project() in External/spirv-tools/CMakeLists.txt

The cmake variable ENABLE_OPT is now no longer changed to preserve the
original state.