Increase TStorageQualifier bitfield width
Fix #3538.

Visual Studio 2022 added a new warning to detect when enumerators can
not be represented within the given bit field width. This warning is
disabled by default but can be enabled using the flag /w15249. This PR
increases the width by one to accommodate the signed maximum value of
EvqLast, which is currently 32.

Perhaps a future improvement would be to use scoped enums; however,
that is more work as the typing is more strict and would require more
changes.