Error for in/out structs containing booleans
The GLSL spec says:
It is a compile-time error to declare a fragment shader input with, or
that contains, any of the following types:
 -a boolean type
 -an opaque type
and similar rules for other shader stages and 'out' variables.

Previously glslang would error for simple 'bool' IO variables, but not
structs/blocks containing them.

This could lead to generating invalid spir-v, violating this validator
rule:

"If OpTypeBool is stored in conjunction with OpVariable using Input or
Output Storage Classes it requires a BuiltIn decoration"
- VUID 7290