Export spv::Parameterize
`spv::Parameterize()` needs to be called before working with SPIR-V.
When using glslang as a library, not through the disassemble or remap
entry-points, `Parameterize` needs to be called explicitly. Therefore
the function needs to be exported.
As `doc.h` is used in spirv-remap, the `visibility.h` header cannot be
included, so redefine the `GLSLANG_EXPORT` macro like it is done in
`SPVRemapper.h`.

Fixes a build with `-DBUILD_SHARED_LIBS=ON`.