Fix spec violations. Simplify tests.
The original fix was accepting invalid GLSL. This reduces the scope of the fix to allow variable declarations to match section 4.2 (Scoping) of the GLSL 4.60 spec: "Each level of structure has its own name space for names given in member declarators; such names need only be unique within that name space."

This means that using a struct name as a member name is valid. The member name hides the type name within the struct.

Updated and simplified test cases to match.

The output of 400.frag's error needed to be updated to reflect the new grammar implementation changes. A struct_declarator can now accept TYPE_NAME in addition to IDENTIFIER, so the parser is now adding TYPE_NAME to the list of expected tokens.