Fix struct types in function parameters
This fixes #4001

The fix for issue #3931 introduced  a regression where struct types could not be used as function parameters when not in the first position. For example:

```
  void test1(B b, int c) {}  // worked
  void test2(int c, B b) {}  // failed with "unexpected IDENTIFIER"
```

The problem was that the declarator list logic was incorrectly being applied to function parameter lists. This adds parens depth tracking so function argument lists are not considered declarator lists (which were forcing IDENTIFIER recognition instead of TYPE).

Added the test from #4001 and also reduced tests from the Tint failures reported in #4001.

Additionally, consolidated some existing tests for clarity.