Allow spec constant matrix use (#3967)
In SPIR-V specification `SPV_KHR_cooperative_matrix`, allows the `use`
of cooperative matrix is a constant instruction. The use in this context
is: how to use the matrix as matrixA, matrixB or matrixAccumulate.

gl_MatrixUse* are constant integer values which can be used for the
MatrixUse template parameter in cooperative matrix types.