Fix NS debug bool type in SSBO/UBO. (#4061)
For GLSL, bool types inside of a buffer or uniform block are translated to uint in SPIR-V. However, the debug info should still represent the type as a bool.

The fix is to construct a bool type in this special case and inject it into the makeMemberDebugType function.

This is more complicated than it should be, because the make*DebugType functions treat their ID argument as a real SPIR-V OpType instea