Add semantic check for cooperative vector loads/stores.
For cooperative vector loads and stores, the buffer arrays cannot be local variables. The compiler was not validating this and when it went to generate debug information, it would hit an assertion because the ID could not be retrieved.

This adds a semantic check that fails compilation when the array is neither in buffer nor shared storage.