Add AST node for variable declarators for better quality debug info (#4074)
At the moment, GLSLANG drops variable declaration info immediately after making the symbol in the symbol table. This leads to difficulty in emitting debug info, presenting that 1) debug info for constant variable is missing 2) debug line of variables points to its first use instead of declaration.

This patch introduces a new TIntermVariableDecl node in the AST, consisting of an TIntermSymbol* declSymbol and an TIntermNode* initNode. declSymbol is an imaginary access to the declared variable, being used to instantiate the basic variable info and debug information at the point of declaration. It doesn't participate AST traversal, such that code generation is still operating on the same AST as before. TIntermNode* initNode is the AST node for the initializer, which used to be emitted directly in the OpSequence. It's now wrapped within a TIntermVariableDecl node. For example,

When a declaration "int ^x;" is seen, a new tree node is inserted to the AST pointing to ^.
When a declaration "int ^x = 42;" is seen, the initializer node in the AST is replace with the new tree node pointing to ^.

Changes:

* Add AST Node for Declaration
We generate a new AST node for each variable declaration (either local/global) when debug info is requested. This means:

- When a declaration "int x;" is seen, a new tree node is inserted to the AST.
- When a declaration "int x = 42;" is seen, the initializer node in the AST is replace with the new tree node.
The newly introduced AST node is a pair (symbol, initializer), where the symbol is the same with the symbol object that's referencing the variable l
- GLSLANG parses block in a different location from regular variables. We have to also handle block instances for more accurate declaration info.
- Only create one declaration node in existence of multiple of them
- Only iterate declaration node once while traversing the AST
- Warp interface block declaration in an OpSequence to ensure it's always processed
- In glslang, builtin redeclaration are not treated as a variable. Ignore them when creating AST for variable declarations.